% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reliabilityStanModels-package.R
\docType{package}
\name{reliabilityStanModels-package}
\alias{reliabilityStanModels-package}
\alias{reliabilityStanModels}
\title{The 'reliabilityStanModels' package.}
\description{
A DESCRIPTION OF THE PACKAGE
}
\references{
Stan Development Team (NA). RStan: the R interface to Stan. R package version 2.32.7. https://mc-stan.org
}
\author{
\strong{Maintainer}: Tim Weber \email{tim.weber@th-deg.de}

}
