% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weibullAndGate.R
\name{weibull_and_gate}
\alias{weibull_and_gate}
\title{The Weibull And Gate Model}
\usage{
weibull_and_gate(t, ...)
}
\arguments{
\item{t}{Numeric vector of input reliability times}

\item{...}{}
}
\value{
An object of class \code{stanfit} returned by \code{rstan::sampling}
}
\description{
The Weibull And Gate Model
}
