% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weibullOrGate.R
\name{weibull_or_gate}
\alias{weibull_or_gate}
\title{The Weibull Or Gate Model}
\usage{
weibull_or_gate(t, ...)
}
\arguments{
\item{t}{Numeric vector of input reliability times}

\item{...}{}
}
\value{
An object of class \code{stanfit} returned by \code{rstan::sampling}
}
\description{
The Weibull Or Gate Model
}
